==> 1-Problem_Framing.ipynb <==
# Frame the Problem and Look at the Big Picture
___

## Define the objective in business terms.

Having chosen a Restaurant, system should recommend similar restaurant based on user ratings.

## How will your solution be used?

Prototype version can be implemented as a web service. Having small web application with a form, where user can input a name of the restaurant. System then returns similar restaurants.

## What are the current workaround solutions/workarounds (if any)?

User can use the yelp website, click around and see what restaurands his friends like etc.

## How should you frame this problem (supervised/unsupervised, online/offline, etc.)?

Supervised learning to train embedding layer using neural networks. Embedding layer is then used to find similar restaurants. It is offline learning. Model is learned in batch mode and deployed for later use. Once in a time it can be retrained.

## How should performance be measured?

Rating is a number between 1 and 5. I use Mean Square Error to measure performance. 

## Is the performance measure aligned with the business objective?

Yes, we want predicted rating to be as close to the actual rating. Because we square errors - bigger difference give bigger errors.

## What would be the minimum performance needed to reach the business objective?

/

## What are comparable problems? Can you reuse experience or tools?

There are several ways to create recommendation system:
 - Collaborative Filtering Recommendations
   -  Based on ratings or reviews of other users
 - Context Based Filtering Recommendations
   -  Based on Attributes of restaurants
 - Friends Network Recommendations
   -  Based on ratings of my friends
 - Location Based Recommendation
   -  Based on user location
>
Sources: (I am reusing / getting inspiration mainly by this sources)
 - https://towardsdatascience.com/building-a-recommendation-system-using-neural-network-embeddings-1ef92e5c80c9
 - https://www.kaggle.com/code/zolboo/recommender-systems-knn-svd-nn-keras
 - https://www.kirenz.com/post/2021-12-11-text-mining-and-sentiment-analysis-with-nltk-and-pandas-in-python/text-mining-and-sentiment-analysis-with-nltk-and-pandas-in-python/
 - https://towardsdatascience.com/how-to-build-a-restaurant-recommendation-system-using-latent-factor-collaborative-filtering-ffe08dd57dca
 - https://stackoverflow.com/questions/45545110/make-pandas-dataframe-apply-use-all-cores 
 - https://stackoverflow.com/questions/18603270/progress-indicator-during-pandas-operations 
 - https://stackoverflow.com/questions/38763007/how-to-use-spacy-lemmatizer-to-get-a-word-into-basic-form 
 - https://stackoverflow.com/questions/22258491/read-a-small-random-sample-from-a-big-csv-file-into-a-python-data-frame
 - https://stats.stackexchange.com/questions/270546/how-does-keras-embedding-layer-work
 - https://github.com/cerlymarco/keras-hypetune
 - etc.

## Is human expertise available?

/

## How would you solve the problem manually?

Possible way, but less exact, prone to errors:
 - I choose a restaurant.
 - Find users that rated this chosen restaurant similarly.
 - See which restaurants they liked the most.
 - Do some statistics and aggregations, like mode, the most liked restaurants by these users.
 - Choose restaurants to recomment from this aggregations.
    

## List the assumptions you (or others) have made so far.

 - I will focus on one city in the USA, to speed up processing.

## Verify assumptions if possible.

/
