==> 2-Getting-Data.ipynb <==
# Get the Data
___

> ### **Note: automate as much of this as possible so you can easily get fresh data.*

## List the data you need and how much you need.

Features needed for the problem of recommendations: user_id, business _id, user_rating, restaurant_name

## Find and document where you can get that data.

CSV:
    https://www.kaggle.com/datasets/yelp-dataset/yelp-dataset/versions/6
<br>
JSON:
    https://www.kaggle.com/datasets/yelp-dataset/yelp-dataset/code?datasetId=10100&sortBy=dateRun&tab=profile&searchQuery=t1&language=Python&types=competitions

## Check how much space it will take.

CSV: over 5.1 GB
<br>
JSON: over 8.6 GB

## Check legal obligations, and get authorization if necessary.

/

## Get access authorizations.

/

## Create a workspace (with enough storage space).

Creating python virtual environment for this project.

## Get the data.

Downloading dataset to the local environment

## Convert the data to a format you can easily manipulate (without changing the data itself).

Using CSV format provided here - https://www.kaggle.com/datasets/yelp-dataset/yelp-dataset/versions/6

## Ensure sensitive information is deleted or protected (e.g. anonymized).

/

## Check the size and type of data (time series, sample, geographical, etc.).

 - CSV has string data types; convering during processing as necessary. 
 - Ids will be label encoded starting from 1 to the number of items;
 - Ratings will be converted to a float.

## Sample a test set, put it aside, and never look at it (no data snooping!).

---

# Libraries
import pandas as pd
import numpy as np

import time

start_time = time.time()

# Read data
#! CAREFULL - Need around 10GB free memory

business = pd.read_csv("../archive\yelp_business.csv")
reviews = pd.read_csv("../archive\yelp_review.csv")

###### --------------------------------------
# Split Business data randomly
business['split'] = np.random.randn(business.shape[0], 1)

# 20% of data will be put aside
msk = np.random.rand(len(business)) >= 0.2

# Applying mask
train_business = business[msk]
test_business = business[~msk]


###### --------------------------------------
# Split reviews data randomly
reviews['split'] = np.random.randn(reviews.shape[0], 1)

# 20% of data will be put aside
msk = np.random.rand(len(reviews)) >= 0.2

# Applying mask
train_reviews = reviews[msk]
test_reviews = reviews[~msk]


###### --------------------------------------
# Store back to CSVs
train_business.to_csv('../archive/train_business.csv') 
test_business.to_csv('../archive/test_business.csv') 
train_reviews.to_csv('../archive/train_reviews.csv') 
test_reviews.to_csv('../archive/test_reviews.csv') 

###### --------------------------------------
# How long it took
end_time = time.time()
print("Took", end_time - start_time, "s")
