==> 4.1-Preparing-Data.ipynb <==
# Prepare the Data
___

 - Work on copies of the data (keep the original dataset intact).
 - Write functions for all data transformations you apply, for five reasons:
    -  So you can easily prepare the data the next time you get a fresh dataset
    -  So you can apply these transformations in future projects
    -  To clean and prepare the test set
    -  To clean and prepare new data instances once your solution is live
    -  To make it easy to treat your preparation choices as hyperparameters
 - Data cleaning:
    -  Fix or remove outliers (optional).
    -  Fill in missing values (e.g., with zero, mean, median...) or drop their rows (or columns).
 - Feature selection (optional):
   -  Drop the attributes that provide no useful information for the task.
 - Feature engineering, where appropriate:
   -  Discretize continuous features.
   -  Decompose features (e.g., categorical, date/time, etc.).
   -  Add promising transformations of features (e.g., log(x), sqrt(x), x^2, etc.).
   -  Aggregate features into promising new features.
 - Feature scaling:
   -  Standardize or normalize features.

---

# Import Libraries
import pandas as pd
import time
import numpy as np

from sklearn.preprocessing import LabelEncoder

# Set printing in non scientific notation
np.set_printoptions(suppress=True)

# Import Data
start_time = time.time()

business = pd.read_csv("../archive/train_business.csv")
reviews = pd.read_csv("../archive/train_reviews.csv", nrows=100000) # sampling by taking first 1M rows, otherwise need a lot of RAM

end_time = time.time()
print("Took", end_time - start_time, "s")


# Businesses in Las Vegas and currently open business
city = business[(business["city"] == "Las Vegas") & (business["is_open"] == 1)]
las_vegas = city[["business_id", "name", "address", "categories", "stars"]]

# Getting just Restaurants
restaurants_only = las_vegas[
    las_vegas["categories"].str.contains("Restaurant.*") == True
].reset_index()

# pull out names and addresses of the restaurants from restaurants_only table
restaurants = restaurants_only[["business_id", "name", "address"]]


# pull out needed columns from review table
df_review = reviews[["user_id", "business_id", "stars", "date"]]


# combine df_review and restaurant table
combined_business_data = pd.merge(df_review, restaurants, on="business_id")
combined_business_data

Label Encoding Users and Restaurants

# Initialize Label Encoder for Users
user_encode = LabelEncoder()

# Add column with Label Encoded Users
combined_business_data["user"] = user_encode.fit_transform(
    combined_business_data["user_id"].values
)

# Initialize Label Encoder for Restaurants
item_encode = LabelEncoder()

# Add column with Label Encoded Restaurants
combined_business_data["business"] = item_encode.fit_transform(
    combined_business_data["business_id"].values
)
combined_business_data.head(2)

# Convert Stars to Float
combined_business_data["stars"] = combined_business_data[
    "stars"
].values.astype(np.float32)

# Check
combined_business_data.info()

combined_business_data.head(2)

# Pickle Data
combined_business_data.to_pickle("combined_business_data.pkl")

# Pickle for Kaggle in protocol 4
combined_business_data.to_pickle("combined_business_data_p4.pkl", protocol=4)
