==> 7-Present-Solution.ipynb <==
# Present Your Solution
___

 - Document what you have done.
 - Create a nice presentation.
 - Make sure you highlight the big picture first.
 - Explain why your solution achieves the business objective.
 - Don't forget to present interesting points you noticed along the way.
 - Describe what worked and what did not.
 - List your assumptions and your system's limitations.
 - Ensure your key findings are communicated through beautiful visualizations or easy-to-remember statements (e.g., "the median income is the number-one predictor of housing prices").


Will be in presentation powerpoint.
